% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\docType{data}
\name{gen}
\alias{gen}
\title{Genotype}
\format{A matrix with 500 rows and 2000 columns:
\describe{
  \item{ID}{individual ID}
  \item{marker}{name of marker}
}}
\usage{
gen
}
\description{
A  matrix with dimension of 500x2000 containing the genotype indicator of
500 simulated individuals of an F2 family generated from the cross of two inbred lines
according to \code{map}.
}
\keyword{datasets}
