% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept.R
\docType{data}
\name{intercept}
\alias{intercept}
\title{Intercept}
\format{A sequence of number containing 30 replicates of number 1.}
\usage{
intercept
}
\description{
Design vector of the intercept for the 30 simulated individuals.
}
\keyword{datasets}
