% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{data}
\name{map}
\alias{map}
\title{Genetic map}
\format{A data frame with 50 rows and 4 variables:
\describe{
  \item{chr}{chromosome}
  \item{marker}{name of marker}
  \item{cm}{genetic distance measured in centi-Morgan}
  \item{effect}{simulated effect assigned to marker in the simulation experiment}
}}
\usage{
map
}
\description{
The genetic map used in the generation of numeric genotype indicators
for 30 simulated individuals of an F2 family generated from the cross of two
inbred lines.
}
\keyword{datasets}
