% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sblas.R
\name{sblas}
\alias{sblas}
\title{Sparse Bayesian learning for QTL mapping and GWAS}
\usage{
sblas(x, y, z, t = -1, max.iter = 200, min.err = 1e-06)
}
\arguments{
\item{x}{a design matrix for fixed effects}

\item{y}{a vector for response variables}

\item{z}{a design matrix for random effects}

\item{t}{a number between [-2,0] to control sparseness of the model,
default is -1.}

\item{max.iter}{maximum number of iterations set by user to stop the program,
default is 200.}

\item{min.err}{minimum threshold of mean squared error of random effects estimated
from the current and the previous iteration to stop the program, default is 1e-6.}
}
\value{
\item{iteration}{a matrix storing intermediate results of each iteration before
the entire program converges, including \cr

"\code{iter}" iteration indicator \cr

"\code{error}" mean squared error of random effects estimated from the current and
the previous iteration \cr

"\code{s2}" estimated variance of residual error \cr

"\eqn{beta[1]\dots beta[p]}" estimates of fixed effects \cr

"\eqn{gamma[1]\dots gamma[m]}" estimates of random effects}

\item{parm}{a vector containing 5 elements: "\code{iter}", "\code{error}",
"\code{s2}", "\code{beta}" and "\code{df}" \cr

"\code{iter}" the number of iterations required by program to stop \cr

"\code{error}" mean square error of random effect estimated from the last iteration
before the program stops \cr

"\code{s2}" estimated variance of residual error \cr

"\code{beta}" estimate of fixed effect \cr

"\code{df}" the effective degree of freedom of total random effects}

\item{blup}{a matrix containing 4 columns: "\code{gamma}", "\code{vg}",
"\code{wald}" and "\code{p_wald}" \cr

"\code{gamma}" estimate of random effect \cr

"\code{vg}" estimated variance of random effect \cr

"\code{wald}" Wald statistic calculated as \eqn{\gamma}/\eqn{\Phi}

"\code{p_wald}" the \emph{p} value of Wald statistic following Chi-squared
distribution with 1 degree of freedom}
}
\description{
The sparse Bayesian learning (SBL) method for QTL mapping and
genome-wide association studies deals with a linear mixed model. This is also
a multiple locus model that includes all markers (random effects) in a single
model and detect significant markers simultaneously. SBL method adopts
coordinate descent algorithm to update parameters by estimating one parameter
at a time conditional on values of all other parameters. The parameter
estimation process requires multiple iteration and the final estimated
parameters take the values when the program converge.
}
\details{
The multiple locus linear mixed model of SBL is
\deqn{y=X\beta+Z\gamma+\epsilon}
where y is a \eqn{n x 1} response variables (observations of the trait);
X is a \eqn{n x p} design matrix for fixed effects; \eqn{\beta} is a \eqn{p x 1}
vector of fixed effect; Z is a \eqn{n x m} genotype indicator matrix;
\eqn{\gamma} is a \eqn{m x 1} vector of marker effects and \eqn{\epsilon} is
a \eqn{n x 1} vector of residual errors with an aassumed
\eqn{\epsilon~N(0,\Sigma)} distribution. Each marker effect, \eqn{\gamma[k]}
for marker \emph{k}, is treated as a random variable following \eqn{N(0,\Phi[k])}
distribution, where \eqn{\Phi[k]} is the prior variance. The estimate of
\eqn{\gamma[k]} is the best linear unbiased prediciton (BLUP). The estimate of
\eqn{\Phi[k]} is miximum likelihood estimate (MLE).
}
\examples{
# Load example data from sbl package
data(gen)
data(phe)
data(intercept)

# Run sblas to perform association study of example data
# set t = 0 to make the model most sparse
fit<-sblas(x=intercept, y=phe, z=gen, t=0)
my.blup<-fit$blup

# set t = -2 to make the model least sparse
fit<-sblas(x=intercept, y=phe, z=gen, t=-2)
my.blup<-fit$blup
}
\author{
Meiyue Wang and Shizhong Xu \cr
Maintainer: Meiyue Wang \email{mwang024@ucr.edu}
}
