% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\docType{data}
\name{gen}
\alias{gen}
\title{Genotype}
\format{A matrix with 30 rows and 50 columns:
\describe{
  \item{ID}{individual ID}
  \item{marker}{name of marker}
}}
\usage{
gen
}
\description{
A  matrix with dimension of 30*50 containing the numeric genotype
indicator for 30 simulated individuals of an F2 family generated from the cross
of two inbred lines according to \code{map} provided in the package.
}
\keyword{datasets}
