% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phe.R
\docType{data}
\name{phe}
\alias{phe}
\title{Phenotype}
\format{A sequence of number with length equaling 500.}
\usage{
phe
}
\description{
Phenotype observations of the 500 simulated individuals.
}
\keyword{datasets}
